% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myTarGetFastStats.R
\name{myTarGetFastStats}
\alias{myTarGetFastStats}
\title{Returns Basic Statistics in Real Time}
\usage{
myTarGetFastStats(
  object_type = "campaigns",
  object_id = NULL,
  auth = NULL,
  login = getOption("rmytarget.login"),
  token_path = myTarTokenPath()
)
}
\arguments{
\item{object_type}{API object type, character value, apply one of campaigns, banners, users}

\item{object_id}{ID of API object (id campaign or any object)}

\item{auth}{R auth object}

\item{login}{Your login, or client name in MyTarget account}

\item{token_path}{Path to directory where you save credential data}
}
\value{
tibble with fast statistics
}
\description{
returns basic statistics on advertising objects in real time, without taking into account the filtering of incorrect traffic. The summary statistics can vary significantly.
}
\examples{
\dontrun{
rt_stat <- myTarGetFastStats()
}
}
