% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myTarGetOfflineConversionsStats.R
\name{myTarGetOfflineConversionsStats}
\alias{myTarGetOfflineConversionsStats}
\title{Returns Statistics on Events Attributed With 'myTarget' Ad Impressions From Offline Conversion Lists for Campaigns}
\usage{
myTarGetOfflineConversionsStats(
  date_from = Sys.Date() - 7,
  date_to = Sys.Date(),
  object_id = NULL,
  auth = NULL,
  login = getOption("rmytarget.login"),
  token_path = myTarTokenPath()
)
}
\arguments{
\item{date_from}{Start date}

\item{date_to}{End date}

\item{object_id}{Campaigns IDs}

\item{auth}{R auth object}

\item{login}{Your login, or client name in MyTarget account}

\item{token_path}{Path to directory where you save credential data}
}
\value{
tibble with offline conversions statistics
}
\description{
Returns Statistics on Events Attributed With 'myTarget' Ad Impressions From Offline Conversion Lists for Campaigns
}
