% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myTarRefreshToken.R
\name{myTarRefreshToken}
\alias{myTarRefreshToken}
\title{Refresh Access Token}
\usage{
myTarRefreshToken(old_auth = NULL, client_id = NULL, client_secret = NULL)
}
\arguments{
\item{old_auth}{Your old auth object. See \code{\link{myTarAuth}}}

\item{client_id}{Your client ID}

\item{client_secret}{Your client secret}
}
\value{
No return value, called for side effects
}
\description{
Refresh Access Token
}
\examples{
\dontrun{
auth <- myTarRefreshToken(old_auth = myTargetAuth,
                         client_id = "xxxx",
                         client_secret = "xxxx.")
}
}
