% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myTarAuth.R
\name{myTarSetLogin}
\alias{myTarSetLogin}
\title{Set MyTarget login}
\usage{
myTarSetLogin(login)
}
\arguments{
\item{login}{Your login, or client name in MyTarget account}
}
\value{
No return value, called for side effects
}
\description{
Set MyTarget login
}
