% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myTarAuth.R
\name{myTarSetTokenPath}
\alias{myTarSetTokenPath}
\title{Set path to auth cache}
\usage{
myTarSetTokenPath(token_path)
}
\arguments{
\item{token_path}{Path to directory where you save credential data}
}
\value{
No return value, called for side effects
}
\description{
Set path to auth cache
}
