% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myTarGetAdList.R
\name{myTarGetAdList}
\alias{myTarGetAdList}
\title{Get ads list from 'MyTarget'}
\usage{
myTarGetAdList(
  auth = NULL,
  login = getOption("rmytarget.login"),
  token_path = myTarTokenPath(),
  request_speed = 1.2
)
}
\arguments{
\item{auth}{MyTarget Autherization R object. See \code{\link{myTarAuth}}}

\item{login}{Account name, used in file name if you save credential}

\item{token_path}{Path to directory where you save credential data.}

\item{request_speed}{Integer, or one of slow, normal, fast. Pause between requests to API}
}
\value{
Data frame with banners list
}
\description{
Get ads list from 'MyTarget'
}
\examples{
\dontrun{
myAds <- myTarGetAdList(login = "my_test_client")
}
}
\seealso{
\href{https://target.my.com/doc/api/ru/resource/Banners}{MyTarget API Documentation}
}
