% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myTarGetClientList.R
\name{myTarGetClientList}
\alias{myTarGetClientList}
\title{Get Agency Clients from 'MyTarget'}
\usage{
myTarGetClientList(
  auth = NULL,
  login = getOption("rmytarget.login"),
  token_path = myTarTokenPath()
)
}
\arguments{
\item{auth}{MyTarget Autherization R object. See \code{\link{myTarAuth}}}

\item{login}{Account name, used in file name if you save credential}

\item{token_path}{Path to directory where you save credential data.}
}
\value{
data frame with agency clients
}
\description{
Get Agency Clients from 'MyTarget'
}
\examples{
\dontrun{
myTarSetLogin("Your Login")
clients <- myTarGetClientList()
}
}
\seealso{
{MyTarget API Documenation}{https://target.my.com/doc/api/detailed/#resource_agency_clients}
}
