% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myTarGetGoalList.R
\name{myTarGetGoalList}
\alias{myTarGetGoalList}
\title{Get goal list from 'MyTarget'}
\usage{
myTarGetGoalList(
  auth = NULL,
  login = getOption("rmytarget.login"),
  token_path = myTarTokenPath()
)
}
\arguments{
\item{auth}{MyTarget Autherization R object. See \code{\link{myTarAuth}}}

\item{login}{Path to directory where you save credential data}

\item{token_path}{Your login, or client name in MyTarget account}
}
\value{
data frame with goals parameters
}
\description{
Get goal list from 'MyTarget'
}
\seealso{
\href{https://target.my.com/doc/api/ru/resource/Goals}{Goals API Docymentation}
}
